+++ plugins/q/src/index.js
@@ -1,11 +1,13 @@
 import data from './data/dataset';
 import qBrushHelper from './brush/q-brush';
-// import qTime from './scales/qTime';
+import numberFormat from './formatter/numberFormat';
+import timeFormat from './formatter/timeFormat';
 
 export default function initialize(picasso) {
   data.util = picasso.data('matrix').util;
   picasso.data('q', data);
-  // picasso.scale('qTime', qTime);
+  picasso.formatter('q-number', numberFormat);
+  picasso.formatter('q-time', timeFormat);
 }
 
 initialize.qBrushHelper = qBrushHelper; // deprecated